Rewrote the character controller to use state machines.
Running and jumping seem to work correctly. Each state is currently a
component, so if that component isn't added to the player object, the
player doesn't have that ability.

There is also a new scene designated for testing, and some debug text
has been added to the screen to show the current state.